requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
gunicorn
pytest
pytest-django
pytest-cov
flake8
python-dotenv